+++ scrapeomatic/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.1.5"
+__version__ = "0.1.6"
