--- scrapeomatic/collectors/tiktok.html
@@ -1,1 +1,0 @@
-<!DOCTYPE html> <html lang="en"> <head>     <link rel="icon" href="data:;base64,=">         <script id="slardar-config" type="application/json"> {     "slardarClient": "SlardarWAF",     "sdkUrl": "https://sf16-website-login.neutral.ttwstatic.com/obj/tiktok_web_login_static/slardar/fe/sdk-web/browser.sg.js",     "bid": "slardar_us_waf",     "pid": "pid",     "pluginPathPrefix": "https://sf16-website-login.neutral.ttwstatic.com/obj/tiktok_web_login_static/slardar/fe/sdk-web/plugins" } </script>     <script src="https://sf16-website-login.neutral.ttwstatic.com/obj/tiktok_web_login_static/slardar/sdk-pre/browser.web.pre.js" crossorigin="anonymous"></script> <script >bazadebezolkohpepadr="58022585"</script><script type="text/javascript" src="https://www.tiktok.com/akam/13/3755a6a"  defer></script></head>  <body> Please wait... <p id="wci" class="_wafchallengeid"></p> <p id="cs" class="eyJ2Ijp7ImEiOiI3aWlNa09sWDQ2eE54Qk04YUZ6a01iRjdLaThUNDJPUGx5ekxDQllpQ3k4PSIsImIiOjE3MDU3OTkxNTUsImMiOiJjLzRNQXM5NmZvSlpBNHFja3JVTkp0S3Zsc3N5QjZsTlR5SnNVb0YzOVRFPSJ9LCJzIjoiSjI0REJEVlBVbXVWUWZvNDFjT1BCcThTVy80Z2JpdVZiSm9TRUV0TGYrWT0ifQ"></p> <script tpye="text/javascript" src="https://sf16-website-login.neutral.ttwstatic.com/obj/tiktok_web_login_static/obj/waf-aiso/dd9799.js"></script> <noscript><img src="https://www.tiktok.com/akam/13/pixel_3755a6a?a=dD1mN2I2NWM0YzJlODhjYWU1NmM2ZWFkZDkwNDczODYwMTcxMTBjZmQ2JmpzPW9mZg==" style="visibility: hidden; position: absolute; left: -999px; top: -999px;" /></noscript></body> </html>
